refactor: Apply clippy suggestions for code quality
- Fixed redundant closure warnings by using direct variant constructors
- Fixed unnecessary_cast warnings in ml_ops.rs (removed 'as u8')
- Fixed ptr_arg warning in ai-cat.rs by using &Path instead of &PathBuf
- Added Path import to ai-cat.rs
- Applied automatic clippy fixes across all binaries

All tests still passing (41 library + 5 integration + 1 doc).

+++ src/ml_ops.rs
@@ -529,7 +529,7 @@ impl FileClassifier {
         // Check for non-printable characters
         let non_printable = content[..sample_size]
             .iter()
-            .filter(|&&b| b < 0x20 && b != b'\t' as u8 && b != b'\n' as u8 && b != b'\r' as u8)
+            .filter(|&&b| b < 0x20 && b != b'\t' && b != b'\n' && b != b'\r')
             .count();
 
         non_printable > sample_size / 20
